refactor: move events in subfolder
