+++ src/Listeners/MessageEventListener.php
@@ -5,9 +5,9 @@ namespace Okaufmann\LaravelNotificationLog\Listeners;
 use Illuminate\Mail\Events\MessageSent;
 use Illuminate\Notifications\Events\NotificationSending;
 use Illuminate\Notifications\Events\NotificationSent;
+use Okaufmann\LaravelNotificationLog\Events\NotificationFailed;
 use Okaufmann\LaravelNotificationLog\Loggers\SentMessageLogger;
 use Okaufmann\LaravelNotificationLog\Loggers\SentNotificationLogger;
-use Okaufmann\LaravelNotificationLog\NotificationFailed;
 
 class MessageEventListener
 {
